fix: Analytics Type 2 Request Body
